xenapi: ensure vdi is not too big when resizing down
Partial fix for bug 1155066

This change adds a check to ensure that when a user chooses to
resize their instance to flavor that is too small,
given their current disk usage, the operation fails gracefully.
It will leave their VM running, but in with the status of Error.

Ideally we can report this back to the user without putting the
VM into an Error state, but this will be a future enhancement.

Change-Id: Ibad568ab3cfb9caaf4fe002572c8cda973d501a7