WIP: prototype filter based scheduler weights
Use filters to generate the list of filter weights. This uses an ordered
list of filters to create a bit field that gives you the list of host
weights.

This also uses the existing weighers system, normalises the output
between 0 and 1, then adds that to the score from the above system.
This is designed purely as a transition method, while all existing
weighers get converted to the filter system. Note filter based weights
always win out any older weighers.

Introduces the new configuration scheduler_soft_filters that allows
users to set the list of filters that should be used for weights.

This relates to the following backlog spec:
https://review.openstack.org/#/c/256323/6/specs/backlog/approved/ordered-filter-scheduler.rst

Change-Id: I4d8ee9f790b5816144e4a6c1b30511977c24d506