WIP: Integrating with unified limits
Alongside the existing quota API this change adds a new limits api that
makes an explicit distinction between:
* instance limits from unified limits (server count, resources classes)
* countable db protection limits (keypairs, server_groups)
* api request db protection limits (metadata_items, injected_files)

The last two could be make the same if the caller was made responsible
for doing the counting. But that kinda defeats the object.

No attempt has been made to adopt context managers to help simplify the
recheck quota logic. It has simply been added alongside the existing
logic, inheriting the same good and bad points of the existing logic.

NOTE:
This uses my own hacked up oslo.limit lib to try and show what sort of
interface might work.
We assume the use of the placement counting logic here, and that all
migration of data that is needed has already been done, etc.

This is just a "sketch", no testing has been done. Just here to share
ideas and help agree a direction.

blueprint unified-limits-stein

Change-Id: If1df93400dcbcb1d3aac0ade80ae5ecf6ce38d11