Better policy unit tests
For each policy file we test with a range of context objects to ensure
we accept and reject the correct sets of contexts.

We test the legacy behaviour of user with the admin role having access
to all system scoped admin APIs, and that any authenticated user has
access to the majoriy of admin_or_owner APIs. The protection really
comes from the DB layer.

In a similar way we test that a system scoped admin token, and a project
scoped token with the member role also work as expected.

This test will need to get slowly updated as we fix various policy bugs,
once we have good functional tests in place to validate subtle behaviour
changes, such as 404 if you don't have access to a server so we do not
leak the existence of that server to an user on a different project.

blueprint policy-defaults-refresh

Change-Id: Ic4bc02f3de63b6fe1007d2a8b87723238a41425d