Add functional test for admin_actions
Rework existing tests to use multiple context, and make a real DB query
with the provided context.

This is to help remove the DB level check and replace it purely with
policy checks in the longer term.

TODO: try add system scope test, and see if it affects the tests
correctly.

Change-Id: Id8b82722ab46c9c97d1f59120b6a2e82e534488c