Ironic: retry when node not available
After a baremetal instance is deleted, and its allocation is removed in placement,
the ironic node might start cleaning. Eventually nova will notice and update the
inventory to be reserved. During this window, a new instance may have already
picked this ironic node.

When that race happens today the build fails with an error:
"Failed to reserve node ..."

This change tries to make sure we instead trigger a re-schedule and try
another node, instead of just aborting the build.

Fixes-Bug: #1974070
Change-Id: Ie5cdc17219c86927ab3769605808cb9d9fa9fa4d