Implementing get_num_instances for ironic virt driver
The default get_num_instances call is provided by the virt interface
which invokes the list_instances method and this method loads the entire
instance via a API call per server by objects.Instance.get_by_uuid.

This is inefficient to get a count of the instances that will already
be held in the node cache.

We instead implement the method and instead call the list_instance_uuids
which results in no API calls provided the cache is present.

This results in a speed up in the sync_power_states periodic task.

Closes-Bug: 2117700
Change-Id: I34a94344a8f68392a28daccc13c76c495eeb9bac
Signed-off-by: Ian Watson <ianwatson92@hotmail.co.uk>